backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==2.5.0
bcrypt==4.1.0
python-jose==3.3.0
pyjwt==2.8.1
python-multipart==0.0.6
